ISO: one gigantic hack to support DVDs with more than 100 minutes of frames (the raw DVD format is probably different but I don't care right now)

+++ Source/RhythmCodex.Lib/FileSystems.Cd/Streamers/CdSectorRange.cs
@@ -21,5 +21,5 @@ public class CdSectorRange(ICdSectorCollection sectors, int start, int count) : 
     public ICdSector this[int index] =>
         sectors[index + start];
 
-    public long Length => count * CdSector.RawSectorSize;
+    public long Length => count * (long)CdSector.RawSectorSize;
 }
